// src/types/types.ts
export type questionType = {
  title: string;
};

export type propertyTypeVariants = {
  var: string;
};

export type carouselObject = {
  src: string;
  title: string;
  desc: string;
  text: string;
};

export type diplomaType = {
  src: string;
  alt: string;
};

export type SocialType = {
  alt: string;
  img: string;
  link: string;
};

export type QuizPropsType = {
  closeModal?: () => void;
};